requirements.txt
pandas==2.1.4
pillow==10.2.0
requests==2.31.0
streamlit==1.30.0